combined skeleton and models
