+++ IGBackend/app/config.py
@@ -1,5 +1,7 @@
 import os
-class Configuration:
-    SECRET_KEY= os.environ.get('SECRET_KEY')
-    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL')
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
+
+class Config:
+  SECRET_KEY=os.environ.get('SECRET_KEY')
+  SQLALCHEMY_TRACK_MODIFICATIONS=False
+  SQLALCHEMY_DATABASE_URI=os.environ.get('DATABASE_URL')
+  SQLALCHEMY_ECHO=True
