+++ IGBackend/app/forms/__init__.py
@@ -1,0 +1,2 @@
+from .login_form import LoginForm
+from .signup_form import SignUpForm
