+++ IGBackend/app/models/__init__.py
@@ -1,0 +1,2 @@
+from .db import db
+from .user import User
