--- IGBackend/app2/forms/__init__.py
@@ -1,2 +1,0 @@
-from .login_form import LoginForm
-from .signup_form import SignUpForm
