--- IGBackend/app2/models/__init__.py
@@ -1,2 +1,0 @@
-from .db import db
-from .user import User
