+++ IGBackend/app_2/config.py
@@ -1,0 +1,5 @@
+import os
+class Configuration:
+    SECRET_KEY= os.environ.get('SECRET_KEY')
+    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL')
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
