+++ IGBackend/app_2/models/Post.py
@@ -12,6 +12,6 @@ class Post(db.Model):
     imagePath = db.Column(db.String(100))
     userId = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
 
-    users = db.relationship("User", back_populates="posts")
-    comments = db.relationship("Comment", back_populates="posts")
-    postLikes = db.relationship("PostLike", back_populates="posts")
+    user = db.relationship("User", back_populates="post")
+    comment = db.relationship("Comment", back_populates="post")
+    postLike = db.relationship("PostLike", back_populates="post")
